# src/main/resources/v3/eric-oss-app-lcm-sdk-openapi.yaml
...
openapi: 3.0.2
info:
  x-api-id: 12f6a604-1712-43a6-a175-467bbcc0327e
  x-audience: external-partner
  description: |
    The App Life Cycle Management (LCM) API controls and monitors the life cycle of the Apps and their instances.
  version: 3.0.0
  title: App Lifecycle Management API
  contact:
    url: https://www.ericsson.com
...
servers:
  - url: https://{eic-host}/app-lifecycle-management/v3
    variables:
      host:
        default: localhost
        description: Change this value to point to your custom host.
paths:
  /apps:
    get:
      summary: Get all Apps
      description: Get a list of all available Apps. This endpoint fetches information about multiple Apps, allowing users to view the details of each App present in the system. Users can use query parameters to filter and refine the list of returned Apps based on specific criteria.
      tags:
        - Apps
      operationId: "getApps"
      parameters:
        - name: name
          in: query
          description: "Filter by name."
          required: false
          schema:
            type: string
        - name: version
          in: query
          description: "Filter by version."
          required: false
          schema:
            type: string
        - name: mode
          in: query
          description: "Filter by mode."
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: "Filter by status."
          required: false
          schema:
            type: string
        - name: type
          in: query
          description: "Filter by type."
          required: false
          schema:
            type: string
        - name: offset
          in: query
          description: |
            To get a paginated list of Apps, use the offset and limit query params to manage the number of Apps to returned.
            The offset specifies the start element for the Apps returned. To get the next page of entries, the API needs to be called with an offset
            parameter that equals the sum of the previous offset value and limit returned in the previous result.
          required: false
          explode: true
          schema:
            type: string
            minimum: 0
          examples:
            offsetExample:
              summary: Example of Offset
              value: /v3/apps?offset=11&limit=10
        - name: limit
          in: query
          description: |
            Limit is used with the offset query param to return a paginated list of Apps. The limit specifies the max number of Apps to return.
          required: false
          style: form
          explode: true
          schema:
            type: string
            minimum: 1
          examples:
            pageLimitExample:
              summary: Example of Limit
              value: /v3/apps?offset=11&limit=10
        - name: Accept
          description: Indicates the response media type accepted by the client.
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK Get Apps response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppItems"
              examples:
                AppItemsResponseOnCreated:
                  $ref: "#/components/examples/AppItemsResponseOnCreated"
                AppItemsResponseOnInitialized:
                  $ref: "#/components/examples/AppItemsResponseOnInitialized"
                AppItemsResponseOnDeInitialized:
                  $ref: "#/components/examples/AppItemsResponseOnDeInitialize"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Apps
      summary: Create an App
      description: Create an App by providing information from the App package. This API is internally utilized by the onboarding microservice for the App creation process.
      operationId: createApp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppRequest'
            examples:
              CreateAppRequest:
                $ref: '#/components/examples/CreateAppRequest'
        required: true
      responses:
        "201":
          description: Created App response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppDetails'
              examples:
                CreateAppResponse:
                  $ref: '#/components/examples/CreateAppResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /apps/{appId}:
    get:
      summary: Get an App based on AppId
      description: Get detailed information about a specific App identified by its unique AppId.
      tags:
        - Apps
      operationId: "getAppById"
      parameters:
        - in: path
          description: "ID of the App to retrieve"
          required: true
          name: appId
          schema:
            type: string
          examples:
            appId:
              summary: Example of App resource identifier
              value: 26471a81-1de4-4ad9-9724-326eefd22230
        - name: Accept
          description: Indicates the response media type accepted by the client.
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppDetails"
              examples:
                GetAppResponse:
                  $ref: "#/components/examples/GetAppResponse"
                GetAppResponseOnInitialize:
                  $ref: "#/components/examples/GetAppResponseOnInitialize"
                GetAppResponseOnInitialized:
                  $ref: "#/components/examples/GetAppResponseOnInitialized"
                GetAppResponseOnDeInitialized:
                  $ref: "#/components/examples/GetAppResponseOnDeInitialized"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete an App based on AppId
      description: Deletes the App identified by its unique AppId from the system. This operation removes the specific App corresponding to the provided AppId, facilitating the removal of app-related information and resources from the system. Use caution as this action is irreversible and permanently removes the designated App from the system.
      tags:
        - Apps
      operationId: "deleteAppById"
      parameters:
        - in: path
          description: "Id of the App to be deleted"
          required: true
          name: appId
          schema:
            type: string
          examples:
            appId:
              summary: Example of App resource identifier
              value: 26471a81-1de4-4ad9-9724-326eefd22230
        - name: Accept
          description: Indicates the response media type accepted by the client.
          in: header
          required: false
          schema:
            type: string
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /apps/{appId}/initialization-actions:
    post:
      tags:
        - Apps
      summary: Initiate an action to initialize or deinitialize the specified App
      description: Initiate a specific action (initialize or deinitialize) for the specified App based on the action specified in the request body.
      operationId: "executeInitializeAction"
      parameters:
        - in: path
          description: "Id of the App to be Initialized or Deinitialized"
          required: true
          name: appId
          schema:
            type: string
          examples:
            appId:
              summary: Example of App resource identifier
              value: 26471a81-1de4-4ad9-9724-326eefd22230
        - name: Accept
          description: Indicates the response media type accepted by the client.
          in: header
          required: false
          schema:
            type: string
        - name: Content-Type
          description: This specifies the media type of the request sent by the client to the server.
          in: header
          required: false
          schema:
            type: string
          example: application/json
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitializeActionRequest'
            examples:
              InitializeAppRequest:
                $ref: '#/components/examples/InitializeAppRequest'
              DeinitializeAppRequest:
                $ref: '#/components/examples/DeinitializeAppRequest'
        required: true
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInitializeOperationResponse'
              examples:
                InitializeAppResponse:
                  $ref: '#/components/examples/InitializeAppResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /apps/{appId}/mode:
    put:
      summary: Enable or disable an App
      description: This endpoint facilitates the enabling or disabling of an App which makes the app available for instantiation.
      operationId: "enableDisableApp"
      tags:
        - Apps
      parameters:
        - in: path
          description: "Id of the App to be Enabled or Disabled"
          name: appId
          schema:
            type: string
          required: true
          examples:
            appId:
              summary: Example of App resource identifier
              value: 26471a81-1de4-4ad9-9724-326eefd22230
        - name: Accept
          description: Indicates the response media type accepted by the client.
          in: header
          required: false
          schema:
            type: string
        - name: Content-Type
          description: This specifies the media type of the request sent by the client to the server.
          in: header
          required: false
          schema:
            type: string
          example: application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnableDisableAppRequest'
            examples:
              EnableAppResponse:
                $ref: "#/components/examples/EnableAppRequest"
              DisableAppResponse:
                $ref: "#/components/examples/DisableAppRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppOperationResponse"
              examples:
                EnableAppResponse:
                  $ref: "#/components/examples/EnableAppResponse"
                DisableAppResponse:
                  $ref: "#/components/examples/DisableAppResponse"
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    CreateAppRequest:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        type:
          type: string
        provider:
          type: string
        components:
          type: array
          items:
            $ref: '#/components/schemas/Component'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
    Component:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        version:
          type: string
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/Artifact'
    Artifact:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        location:
          type: string
    Permission:
      type: object
      properties:
        resource:
          type: string
        scope:
          type: string
    Role:
      type: object
      properties:
        name:
          type: string
    InitializeActionRequest:
      type: object
      properties:
        action:
          description: The type of initializations to be executed, for example, initialize or deinitialize App
          type: string
    EnableDisableAppRequest:
      type: object
      properties:
        mode:
          $ref: '#/components/schemas/AppMode'
    AppMode:
      description: App mode
      type: string
      enum:
        - ENABLED
        - DISABLED
...
    AppItems:
      type: object
      properties:
        items:
          description: The list of all Apps
          type: array
          items:
            $ref: "#/components/schemas/AppDetails"
    AppDetails:
      required:
        - id
        - type
        - provider
        - name
        - version
        - mode
        - status
        - createdAt
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        provider:
          type: string
        name:
          type: string
        version:
          type: string
        mode:
          $ref: '#/components/schemas/AppMode'
        status:
          $ref: '#/components/schemas/AppStatus'
        createdAt:
          type: string
        components:
          type: array
          items:
            $ref: '#/components/schemas/Component'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        self:
          $ref: '#/components/schemas/Href'
    AppStatus:
      type: string
      enum:
        - CREATED
        - DELETED
        - DELETE_ERROR
        - INITIALIZING
        - INITIALIZED
        - INITIALIZE_ERROR
        - DEINITIALIZING
        - DEINITIALIZED
        - DEINITIALIZE_ERROR
      description: App status
    Event:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        detail:
          type: string
        createdAt:
          type: string
    Href:
      type: object
      properties:
        href:
          type: string
    AppOperationResponse:
      type: object
      properties:
        mode:
          $ref: '#/components/schemas/AppMode'
        app:
          properties:
            id:
              type: string
            href:
              type: string
    AppInitializeOperationResponse:
      type: object
      properties:
        app:
          properties:
            status:
              $ref: '#/components/schemas/AppStatus'
            id:
              type: string
            href:
              type: string
...
    ProblemDetails:
      required:
        - title
        - status
        - detail
      type: object
      properties:
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code generated by the server
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            BadRequest:
              $ref: '#/components/examples/BadRequest'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            NotFound:
              $ref: '#/components/examples/NotFound'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            InternalServerError:
              $ref: '#/components/examples/InternalServerError'
  examples:
...
    CreateAppRequest:
      summary: Create App request body
      value:
        name: eric-oss-hello-world-multiple-microservices-go-app
        version: 1.1.1
        type: rApp
        provider: Ericsson
        components:
          - type: Microservice
            name: eric-oss-hello-metrics-go-app
            version: 1.1.1
            artifacts:
              - type: HELM
                name: eric-oss-hello-world-go-app
                location: 26471a81-1de4-4ad9-9724-326eefd22230/eric-oss-hello-world-go-app
              - type: IMAGE
                name: docker.tar
                location: 26471a81-1de4-4ad9-9724-326eefd22230/docker
        permissions:
          - resource: nginx
            scope: foo
          - resource: mongoDB
            scope: goodbyeWorld
        roles:
          - name: admin
    InitializeAppRequest:
      summary: Initialize App Request
      value:
        action: "INITIALIZE"
    DeinitializeAppRequest:
      summary: Sample Deinitialize App Request
      value:
        action: "DEINITIALIZE"
...
    EnableAppRequest:
      summary: Enable App Request
      value:
        mode: "ENABLED"
    DisableAppRequest:
      summary: Disable App Request
      value:
        mode: "DISABLED"
...
    CreateAppResponse:
      summary: Create App request
      value:
        id: 26471a81-1de4-4ad9-9724-326eefd22230
        name: eric-oss-hello-world-multiple-microservices-go-app
        version: 1.1.1
        mode: DISABLED
        status: CREATED
        createdAt: "2023-04-06T00:04:16.711+00:00"
        type: rApp
        provider: Ericsson
        components:
          - type: Microservice
            name: eric-oss-hello-metrics-go-app
            version: 1.1.1
            artifacts:
              - type: HELM
                name: eric-oss-hello-world-go-app
                location: 26471a81-1de4-4ad9-9724-326eefd22230/eric-oss-hello-world-go-app
              - type: IMAGE
                name: docker.tar
                location: 26471a81-1de4-4ad9-9724-326eefd22230/docker
        permissions:
          - resource: nginx
            scope: foo
          - resource: mongoDB
            scope: goodbyeWorld
        roles:
          - name: admin
        events: [ ]
        self:
          href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
    InitializeAppResponse:
      summary: Initialize App Response
      value:
        app:
          id: 26471a81-1de4-4ad9-9724-326eefd22230
          status: "INITIALIZING"
          href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
    EnableAppResponse:
      summary: Enable App Response
      value:
        mode: "ENABLED"
        app:
          id: 26471a81-1de4-4ad9-9724-326eefd22230
          href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
    DisableAppResponse:
      summary: Disable App Response
      value:
        mode: "DISABLED"
        app:
          id: 26471a81-1de4-4ad9-9724-326eefd22230
          href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
...
    BadRequest:
      value:
        title: Bad Request
        status: 400
        detail: The message body is missing a required field.
    NotFound:
      value:
        title: Not Found
        status: 404
        detail: The requested URL is not found.
    InternalServerError:
      value:
        title: Internal Server Error
        status: 500
        detail: Internal Server Error Occurs.
...
    AppItemsResponseOnCreated:
      summary: Get App Response on CREATED
      value:
        items:
          - id: 26471a81-1de4-4ad9-9724-326eefd22230
            type: rApp
            name: eric-oss-hello-world-multiple-microservices-go-app
            version: 1.1.1
            mode: DISABLED
            status: CREATED
            createdAt: "2024-02-20 10:51:45.427503"
            components:
              - type: Microservice
                name: eric-oss-hello-metrics-go-app
                version: 1.1.1
                artifacts:
                  - type: HELM
                    name: eric-oss-hello-world-go-app
                    location: 26471a81-1de4-4ad9-9724-326eefd22230/eric-oss-hello-world-go-app
                  - type: IMAGE
                    name: docker.tar
                    location: 26471a81-1de4-4ad9-9724-326eefd22230/docker
            permissions:
              - resource: nginx
                scope: foo
              - resource: mongoDB
                scope: goodbyeWorld
            roles:
              - name: admin
            events: [ ]
            self:
              href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
    AppItemsResponseOnInitialized:
      summary: Get App Response on INITIALIZED
      value:
        items:
          - id: 26471a81-1de4-4ad9-9724-326eefd22230
            type: rApp
            name: eric-oss-hello-world-multiple-microservices-go-app
            version: 1.1.1
            mode: ENABLED
            status: INITIALIZED
            createdAt: "2024-02-20 10:51:45.427503"
            components:
              - type: Microservice
                name: eric-oss-hello-metrics-go-app
                version: 1.1.1
                artifacts:
                  - type: HELM
                    name: eric-oss-hello-world-go-app
                    location: 26471a81-1de4-4ad9-9724-326eefd22230/eric-oss-hello-world-go-app
                  - type: IMAGE
                    name: docker.tar
                    location: 26471a81-1de4-4ad9-9724-326eefd22230/docker
            permissions:
              - resource: nginx
                scope: foo
              - resource: mongoDB
                scope: goodbyeWorld
            roles:
              - name: admin
            events: [ ]
            self:
              href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
    AppItemsResponseOnDeInitialize:
      summary: Get App Response on DEINITIALIZE
      value:
        items:
          - id: 26471a81-1de4-4ad9-9724-326eefd22230
            type: rApp
            name: eric-oss-hello-world-multiple-microservices-go-app
            version: 1.1.1
            mode: DISABLED
            status: CREATED
            createdAt: "2024-02-20 10:51:45.427503"
            components:
              - type: Microservice
                name: eric-oss-hello-metrics-go-app
                version: 1.1.1
                artifacts:
                  - type: HELM
                    name: eric-oss-hello-world-go-app
                    location: 26471a81-1de4-4ad9-9724-326eefd22230/eric-oss-hello-world-go-app
                  - type: IMAGE
                    name: docker.tar
                    location: 26471a81-1de4-4ad9-9724-326eefd22230/docker
            permissions:
              - resource: nginx
                scope: foo
              - resource: mongoDB
                scope: goodbyeWorld
            roles:
              - name: admin
            events: [ ]
            self:
              href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
    GetAppResponse:
      summary: Get App Response
      value:
        id: 26471a81-1de4-4ad9-9724-326eefd22230
        name: eric-oss-hello-world-multiple-microservices-go-app
        version: 1.1.1
        mode: DISABLED
        status: CREATED
        createdAt: "2023-04-06T00:04:16.711+00:00"
        type: rApp
        components:
          - type: Microservice
            name: eric-oss-hello-metrics-go-app
            version: 1.1.1
            artifacts:
              - type: HELM
                name: eric-oss-hello-world-go-app
                location: 26471a81-1de4-4ad9-9724-326eefd22230/eric-oss-hello-world-go-app
              - type: IMAGE
                name: docker.tar
                location: 26471a81-1de4-4ad9-9724-326eefd22230/docker
        permissions:
          - resource: nginx
            scope: foo
          - resource: mongoDB
            scope: goodbyeWorld
        roles:
          - name: admin
        events: [ ]
        self:
          href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
    GetAppResponseOnInitialize:
      summary: Get App when INITIALIZED is triggered
      value:
        id: 26471a81-1de4-4ad9-9724-326eefd22230
        name: eric-oss-hello-world-multiple-microservices-go-app
        version: 1.1.1
        mode: DISABLED
        status: INITIALIZING
        createdAt: "2023-04-06T00:04:16.711+00:00"
        type: rApp
        components:
          - type: Microservice
            name: eric-oss-hello-metrics-go-app
            version: 1.1.1
            artifacts:
              - type: HELM
                name: eric-oss-hello-world-go-app
                location: 26471a81-1de4-4ad9-9724-326eefd22230/eric-oss-hello-world-go-app
              - type: IMAGE
                name: docker.tar
                location: 26471a81-1de4-4ad9-9724-326eefd22230/docker
        permissions:
          - resource: nginx
            scope: foo
          - resource: mongoDB
            scope: goodbyeWorld
        roles:
          - name: admin
        events: [ ]
        self:
          href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
    GetAppResponseOnInitialized:
      summary: Get App when INITIALIZED is successful
      value:
        id: 26471a81-1de4-4ad9-9724-326eefd22230
        name: eric-oss-hello-world-multiple-microservices-go-app
        version: 1.1.1
        mode: DISABLED
        status: INITIALIZED
        createdAt: "2023-04-06T00:04:16.711+00:00"
        type: rApp
        components:
          - type: Microservice
            name: eric-oss-hello-metrics-go-app
            version: 1.1.1
            artifacts:
              - type: HELM
                name: eric-oss-hello-world-go-app
                location: 26471a81-1de4-4ad9-9724-326eefd22230/eric-oss-hello-world-go-app
              - type: IMAGE
                name: docker.tar
                location: 26471a81-1de4-4ad9-9724-326eefd22230/docker
        permissions:
          - resource: nginx
            scope: foo
          - resource: mongoDB
            scope: goodbyeWorld
        roles:
          - name: admin
        events: [ ]
        self:
          href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
    GetAppResponseOnDeInitialized:
      summary: Get App when DEINITIALIZE is triggered
      value:
        id: 26471a81-1de4-4ad9-9724-326eefd22230
        name: eric-oss-hello-world-multiple-microservices-go-app
        version: 1.1.1
        mode: ENABLED
        status: DEINITIALIZING
        createdAt: "2023-04-06T00:04:16.711+00:00"
        type: rApp
        components:
          - type: Microservice
            name: eric-oss-hello-metrics-go-app
            version: 1.1.1
            artifacts:
              - type: HELM
                name: eric-oss-hello-world-go-app
                location: 26471a81-1de4-4ad9-9724-326eefd22230/eric-oss-hello-world-go-app
              - type: IMAGE
                name: docker.tar
                location: 26471a81-1de4-4ad9-9724-326eefd22230/docker
        permissions:
          - resource: nginx
            scope: foo
          - resource: mongoDB
            scope: goodbyeWorld
        roles:
          - name: admin
        events: [ ]
        self:
          href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
...